Tim's Upload
Partially got the nav menu icon to work (we should probably just keep using the hamburger icon)
Fixed Work Page so modal's not stuck inside small description
Styled the contact page a bit


My recommendations
- Let's keep things basic for now, I admire your wanting to add a lot of different interactions and using a lot of motion, but we've straddled the line of what's really the main focus of web design and into Web Development.
- turn off Vanilla Tilt JS - It's pretty advanced and causing issues when it doesn't work.
- make sure to get all of the rest of your images in the work page.

Literally the last things that will need done is to ensure things will work on a smaller screen sizes and that's about it. Carousels on small devices are notoriously difficult so we may make some compromises on functionality.